+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased](https://github.com/sanette/bogue/compare/20250815...HEAD)
 
+### Added
+
+- `I18n.make_context` is now public.
+
 ## [20250815](https://github.com/sanette/bogue/compare/20250224...20250815)
 
 ### Added